task config generate
