+++ config/.env.example
@@ -3,6 +3,25 @@ RIVERBOAT_RIVER_DATABASEHOST="postgres://postgres:password@0.0.0.0:5432/jobs?ssl
 RIVERBOAT_RIVER_QUEUES=""
 RIVERBOAT_RIVER_WORKERS_OPENLANECONFIG_OPENLANEAPIHOST=""
 RIVERBOAT_RIVER_WORKERS_OPENLANECONFIG_OPENLANEAPITOKEN=""
+RIVERBOAT_RIVER_WORKERS_EMAILCONFIG_COMPANYNAME=""
+RIVERBOAT_RIVER_WORKERS_EMAILCONFIG_COMPANYADDRESS=""
+RIVERBOAT_RIVER_WORKERS_EMAILCONFIG_CORPORATION=""
+RIVERBOAT_RIVER_WORKERS_EMAILCONFIG_YEAR=""
+RIVERBOAT_RIVER_WORKERS_EMAILCONFIG_FROMEMAIL=""
+RIVERBOAT_RIVER_WORKERS_EMAILCONFIG_SUPPORTEMAIL=""
+RIVERBOAT_RIVER_WORKERS_EMAILCONFIG_QUESTIONNAIREEMAIL=""
+RIVERBOAT_RIVER_WORKERS_EMAILCONFIG_LOGOURL=""
+RIVERBOAT_RIVER_WORKERS_EMAILCONFIG_URLS_ROOT=""
+RIVERBOAT_RIVER_WORKERS_EMAILCONFIG_URLS_PRODUCT=""
+RIVERBOAT_RIVER_WORKERS_EMAILCONFIG_URLS_DOCS=""
+RIVERBOAT_RIVER_WORKERS_EMAILCONFIG_URLS_VERIFY=""
+RIVERBOAT_RIVER_WORKERS_EMAILCONFIG_URLS_INVITE=""
+RIVERBOAT_RIVER_WORKERS_EMAILCONFIG_URLS_RESET=""
+RIVERBOAT_RIVER_WORKERS_EMAILCONFIG_URLS_VERIFYSUBSCRIBER=""
+RIVERBOAT_RIVER_WORKERS_EMAILCONFIG_URLS_VERIFYBILLING=""
+RIVERBOAT_RIVER_WORKERS_EMAILCONFIG_URLS_BILLING=""
+RIVERBOAT_RIVER_WORKERS_EMAILCONFIG_URLS_QUESTIONNAIRE=""
+RIVERBOAT_RIVER_WORKERS_EMAILCONFIG_TEMPLATESPATH=""
 RIVERBOAT_RIVER_WORKERS_EMAILWORKER_CONFIG_ENABLED="false"
 RIVERBOAT_RIVER_WORKERS_EMAILWORKER_CONFIG_DEVMODE="true"
 RIVERBOAT_RIVER_WORKERS_EMAILWORKER_CONFIG_TESTDIR="fixtures/email"
@@ -24,6 +43,36 @@ RIVERBOAT_RIVER_WORKERS_DELETEEXPORTCONTENTWORKER_CONFIG_CUTOFFDURATION=""
 RIVERBOAT_RIVER_WORKERS_SLACKWORKER_CONFIG_ENABLED="false"
 RIVERBOAT_RIVER_WORKERS_SLACKWORKER_CONFIG_DEVMODE="true"
 RIVERBOAT_RIVER_WORKERS_SLACKWORKER_CONFIG_TOKEN=""
+RIVERBOAT_RIVER_WORKERS_ORGANIZATIONDELETIONREMINDERWORKER_CONFIG_OPENLANEAPIHOST=""
+RIVERBOAT_RIVER_WORKERS_ORGANIZATIONDELETIONREMINDERWORKER_CONFIG_OPENLANEAPITOKEN=""
+RIVERBOAT_RIVER_WORKERS_ORGANIZATIONDELETIONREMINDERWORKER_CONFIG_PAYMENTMETHODINTERVAL=""
+RIVERBOAT_RIVER_WORKERS_ORGANIZATIONDELETIONREMINDERWORKER_CONFIG_DELETIONDAYS=""
+RIVERBOAT_RIVER_WORKERS_ORGANIZATIONDELETIONREMINDERWORKER_CONFIG_ENABLED=""
+RIVERBOAT_RIVER_WORKERS_ORGANIZATIONDELETIONREMINDERWORKER_CONFIG_EMAIL_ENABLED=""
+RIVERBOAT_RIVER_WORKERS_ORGANIZATIONDELETIONREMINDERWORKER_CONFIG_EMAIL_CONFIG_COMPANYNAME=""
+RIVERBOAT_RIVER_WORKERS_ORGANIZATIONDELETIONREMINDERWORKER_CONFIG_EMAIL_CONFIG_COMPANYADDRESS=""
+RIVERBOAT_RIVER_WORKERS_ORGANIZATIONDELETIONREMINDERWORKER_CONFIG_EMAIL_CONFIG_CORPORATION=""
+RIVERBOAT_RIVER_WORKERS_ORGANIZATIONDELETIONREMINDERWORKER_CONFIG_EMAIL_CONFIG_YEAR=""
+RIVERBOAT_RIVER_WORKERS_ORGANIZATIONDELETIONREMINDERWORKER_CONFIG_EMAIL_CONFIG_FROMEMAIL=""
+RIVERBOAT_RIVER_WORKERS_ORGANIZATIONDELETIONREMINDERWORKER_CONFIG_EMAIL_CONFIG_SUPPORTEMAIL=""
+RIVERBOAT_RIVER_WORKERS_ORGANIZATIONDELETIONREMINDERWORKER_CONFIG_EMAIL_CONFIG_QUESTIONNAIREEMAIL=""
+RIVERBOAT_RIVER_WORKERS_ORGANIZATIONDELETIONREMINDERWORKER_CONFIG_EMAIL_CONFIG_LOGOURL=""
+RIVERBOAT_RIVER_WORKERS_ORGANIZATIONDELETIONREMINDERWORKER_CONFIG_EMAIL_CONFIG_URLS_ROOT=""
+RIVERBOAT_RIVER_WORKERS_ORGANIZATIONDELETIONREMINDERWORKER_CONFIG_EMAIL_CONFIG_URLS_PRODUCT=""
+RIVERBOAT_RIVER_WORKERS_ORGANIZATIONDELETIONREMINDERWORKER_CONFIG_EMAIL_CONFIG_URLS_DOCS=""
+RIVERBOAT_RIVER_WORKERS_ORGANIZATIONDELETIONREMINDERWORKER_CONFIG_EMAIL_CONFIG_URLS_VERIFY=""
+RIVERBOAT_RIVER_WORKERS_ORGANIZATIONDELETIONREMINDERWORKER_CONFIG_EMAIL_CONFIG_URLS_INVITE=""
+RIVERBOAT_RIVER_WORKERS_ORGANIZATIONDELETIONREMINDERWORKER_CONFIG_EMAIL_CONFIG_URLS_RESET=""
+RIVERBOAT_RIVER_WORKERS_ORGANIZATIONDELETIONREMINDERWORKER_CONFIG_EMAIL_CONFIG_URLS_VERIFYSUBSCRIBER=""
+RIVERBOAT_RIVER_WORKERS_ORGANIZATIONDELETIONREMINDERWORKER_CONFIG_EMAIL_CONFIG_URLS_VERIFYBILLING=""
+RIVERBOAT_RIVER_WORKERS_ORGANIZATIONDELETIONREMINDERWORKER_CONFIG_EMAIL_CONFIG_URLS_BILLING=""
+RIVERBOAT_RIVER_WORKERS_ORGANIZATIONDELETIONREMINDERWORKER_CONFIG_EMAIL_CONFIG_URLS_QUESTIONNAIRE=""
+RIVERBOAT_RIVER_WORKERS_ORGANIZATIONDELETIONREMINDERWORKER_CONFIG_EMAIL_CONFIG_TEMPLATESPATH=""
+RIVERBOAT_RIVER_WORKERS_ORGANIZATIONDELETIONWORKER_CONFIG_OPENLANEAPIHOST=""
+RIVERBOAT_RIVER_WORKERS_ORGANIZATIONDELETIONWORKER_CONFIG_OPENLANEAPITOKEN=""
+RIVERBOAT_RIVER_WORKERS_ORGANIZATIONDELETIONWORKER_CONFIG_RUNINTERVAL=""
+RIVERBOAT_RIVER_WORKERS_ORGANIZATIONDELETIONWORKER_CONFIG_MAXDELETESPERRUN=""
+RIVERBOAT_RIVER_WORKERS_ORGANIZATIONDELETIONWORKER_CONFIG_ENABLED=""
 RIVERBOAT_RIVER_TRUSTCENTERWORKERS_OPENLANECONFIG_OPENLANEAPIHOST=""
 RIVERBOAT_RIVER_TRUSTCENTERWORKERS_OPENLANECONFIG_OPENLANEAPITOKEN=""
 RIVERBOAT_RIVER_DEFAULTMAXRETRIES="10"
