+++ go.mod
@@ -36,6 +36,7 @@ require (
 	github.com/stoewer/go-strcase v1.3.1
 	github.com/stretchr/testify v1.11.1
 	github.com/theopenlane/core/common v1.0.20
+	github.com/theopenlane/corejobs v0.1.21
 	github.com/theopenlane/dbx v0.1.3
 	github.com/theopenlane/emailtemplates v0.3.7
 	github.com/theopenlane/go-client v0.9.4
@@ -108,7 +109,6 @@ require (
 	github.com/spf13/pflag v1.0.10 // indirect
 	github.com/stretchr/objx v0.5.3 // indirect
 	github.com/theopenlane/core v1.16.8 // indirect
-	github.com/theopenlane/corejobs v0.1.21 // indirect
 	github.com/theopenlane/echox v0.3.0 // indirect
 	github.com/tidwall/gjson v1.18.0 // indirect
 	github.com/tidwall/match v1.2.0 // indirect
