+++ internal/river/workers.go
@@ -4,12 +4,13 @@ import (
 	"github.com/riverqueue/river"
 	"github.com/rs/zerolog/log"
 
+	"github.com/theopenlane/corejobs"
 	"github.com/theopenlane/riverboat/pkg/jobs"
 	"github.com/theopenlane/riverboat/pkg/riverqueue"
 )
 
 // createWorkers creates a new workers instance
-func createWorkers(w Workers, _ *riverqueue.Client) (*river.Workers, error) {
+func createWorkers(w Workers, insertOnlyClient *riverqueue.Client) (*river.Workers, error) {
 	// create workers
 	workers := river.NewWorkers()
 
@@ -35,6 +36,10 @@ func createWorkers(w Workers, _ *riverqueue.Client) (*river.Workers, error) {
 	}
 
 	if err := createExportWorkers(w, workers); err != nil {
+		return nil, err
+	}
+
+	if err := createOrganizationDeletionWorkers(w, insertOnlyClient, workers); err != nil {
 		return nil, err
 	}
 
@@ -66,6 +71,47 @@ func createExportWorkers(w Workers, workers *river.Workers) error {
 		}
 
 		log.Info().Msg("worker enabled: delete export content")
+	}
+
+	return nil
+}
+
+func setAndValidateOpenlaneConfigDefaults(input *corejobs.OpenlaneConfig, target jobs.OpenlaneConfig) error {
+	input.SetAPIHost(target.GetAPIHost())
+	input.SetAPIToken(target.GetAPIToken())
+
+	return input.Validate()
+}
+
+func createOrganizationDeletionWorkers(w Workers, insertOnlyClient *riverqueue.Client, workers *river.Workers) error {
+	if w.OrganizationDeletionReminderWorker.Config.Enabled {
+		if err := setAndValidateOpenlaneConfigDefaults(&w.OrganizationDeletionReminderWorker.Config.OpenlaneConfig, w.OpenlaneConfig); err != nil {
+			log.Error().Err(err).Msg("failed to set and validate openlane config defaults for organization deletion reminder worker")
+			return err
+		}
+
+		w.EmailConfig.SetDefaultsIfUnset(&w.OrganizationDeletionReminderWorker.Config.Email.Config)
+
+		w.OrganizationDeletionReminderWorker.WithRiverClient(insertOnlyClient)
+
+		if err := river.AddWorkerSafely(workers, &w.OrganizationDeletionReminderWorker); err != nil {
+			return err
+		}
+
+		log.Info().Msg("worker enabled: organization deletion reminder")
+	}
+
+	if w.OrganizationDeletionWorker.Config.Enabled {
+		if err := setAndValidateOpenlaneConfigDefaults(&w.OrganizationDeletionWorker.Config.OpenlaneConfig, w.OpenlaneConfig); err != nil {
+			log.Error().Err(err).Msg("failed to set and validate openlane config defaults for organization deletion worker")
+			return err
+		}
+
+		if err := river.AddWorkerSafely(workers, &w.OrganizationDeletionWorker); err != nil {
+			return err
+		}
+
+		log.Info().Msg("worker enabled: organization deletion")
 	}
 
 	return nil
